requirements.txt
streamlit 
replicate 
imgurpython
python-dotenv

